requirements.txt
PyMuPDF==1.23.8
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
pytesseract==0.3.10
SQLAlchemy==2.0.23
tqdm==4.66.1
# Optional - uncomment if Visual C++ Build Tools installed:
# paddlepaddle==2.5.2
# paddleocr==2.7.3
